refactor: utils & ref message
