+++ web/src/beta/lib/core/engines/Cesium/Feature/Tileset/hooks.ts
@@ -44,7 +44,7 @@ import {
   convertCesium3DTileFeatureProperties,
   lookupFeatures,
   translationWithClamping,
-} from "../../utils";
+} from "../../utils/utils";
 import { useContext } from "../context";
 import {
   usePick,
