+++ web/src/beta/lib/core/engines/Cesium/utils/polygon.ts
@@ -2,7 +2,29 @@
 
 import { Cartesian3, type PolygonHierarchy } from "@cesium/engine";
 import unkinkPolygon from "@turf/unkink-polygon";
-import { type MultiPolygon, type Polygon, type Position } from "geojson";
+import type { LineString, MultiPolygon, Polygon, Position } from "geojson";
+
+export function isNotNullish<T>(value: T | null | undefined): value is T {
+  return value != null;
+}
+
+function coordinatesToPositionsArray(coordinates: readonly Position[][]): Cartesian3[][] {
+  return coordinates.map(coordinates =>
+    coordinates.map(([x, y]) => Cartesian3.fromDegrees(x, y, 0)),
+  );
+}
+
+export function convertGeometryToPositionsArray(
+  geometry: LineString | Polygon | MultiPolygon,
+): Cartesian3[][] {
+  return (
+    geometry.type === "LineString"
+      ? coordinatesToPositionsArray([geometry.coordinates])
+      : geometry.type === "Polygon"
+      ? coordinatesToPositionsArray(geometry.coordinates)
+      : geometry.coordinates.flatMap(coordinates => coordinatesToPositionsArray(coordinates))
+  ).filter(({ length }) => length > 0);
+}
 
 function coordinatesToHierarchy(coordinates: readonly Position[][]): PolygonHierarchy | undefined {
   return coordinates.length > 0
@@ -23,8 +45,4 @@ export function convertPolygonToHierarchyArray(
   return polygons
     .map(polygon => coordinatesToHierarchy(polygon.geometry.coordinates))
     .filter(isNotNullish);
-}
-
-export function isNotNullish<T>(value: T | null | undefined): value is T {
-  return value != null;
 }
